requirements.txt
pytest==7.2.1
allure-pytest==2.12.0
